refactor: modularize main responsibilities
